persist var lib

+++ exe/setup_storage.sh
@@ -11,10 +11,12 @@ for path in ${paths[@]}; do
 	ln -sv $persist$path $path
 done
 
-mkdir -p /var/lib/kubelet/pki
-cp -rv $persist/var/lib/kubelet/pki/kubelet* /var/lib/kubelet/pki/
-rm -fv /var/lib/kubelet/pki/kubelet-client-current.pem
-ln -sv /var/lib/kubelet/pki/$(ls /var/lib/kubelet/pki/ | tail -3 | head -1) /var/lib/kubelet/pki/kubelet-client-current.pem 
+cp -rv $persist/var/lib/ /var/lib/
+
+if [ -d /var/lib/kubelet/pki/ ] && [ ! -z "$(ls /var/lib/kubelet/pki)" ]; then
+	rm -fv /var/lib/kubelet/pki/kubelet-client-current.pem
+	ln -sv /var/lib/kubelet/pki/$(ls /var/lib/kubelet/pki/ | tail -3 | head -1) /var/lib/kubelet/pki/kubelet-client-current.pem
+fi
 
 
 # This was moved to mkkubelet.sh
